Use only the BrowsingData API to clear history.
This may make more sense. 

The function clearBrowsingData(), which is called under three circumstances listed below, does clear all browsing history - but I personally consider that to be more inline with 'disabling' history anyway.

The circumstances under which the function is called are:

- On first install of the extension.
- When an item is added to history.
- When a tab is closed.

I will add an explanation as to why these are required in DOCUMENTATION.md.